Move deno_std to a more convenient location. (#3057)
js/deps/https/deno.land/std -> js/std
