+++ cli/tests/034_onload/nest_imported.ts
@@ -1,4 +1,4 @@
-import { assert } from "../../../js/deps/https/deno.land/std/testing/asserts.ts";
+import { assert } from "../../../std/testing/asserts.ts";
 
 const handler = (e: Event): void => {
   assert(!e.cancelable);
