+++ tools/benchmark.py
@@ -211,8 +211,8 @@ def run_http(build_dir, new_data):
 
 def bundle_benchmark(deno_exe):
     bundles = {
-        "file_server": "./js/deps/https/deno.land/std/http/file_server.ts",
-        "gist": "./js/deps/https/deno.land/std/examples/gist.ts",
+        "file_server": "./std/http/file_server.ts",
+        "gist": "./std/examples/gist.ts",
     }
 
     sizes = {}
