+++ website/app_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 
 import { test, assertEquals } from "../js/test_util.ts";
-import { runIfMain } from "../js/deps/https/deno.land/std/testing/mod.ts";
+import { runIfMain } from "../std/testing/mod.ts";
 import {
   BenchmarkRun,
   createBinarySizeColumns,
